normalize the cross correlations by the auto correlations
This essentially removes the sawtooth on a per-antenna basis. This is
necessary because it seems like the fiber antennas have a different
pattern of gain fluctuations from the core antennas. Presumably this is
due to temperature sensitivity in the laser receiver boards. Note that
this may introduce additional noise into the cross correlations but this
noise will be lower in amplitude than the sawtooth pattern itself, so
this is a net win.