better implementation of cleaning
This is much closer to what I have sitting in the old-code/
sub-directory. Essentially we're only doing major iterations. We get the
position from a centroid, solve for the PSF at that location and then
scale and remove that component from the maps.

We don't putz around with any image-space algorithms because that
involes either working with Healpix maps (where the geometry of the
pixels is not intuitively obvious) or we interpolate the Healpix maps
onto a rectangular grid (in which case we somehow need to undo the
interpolation at a later date).

An added bonus is that this works a lot better.

Finally, it takes about 8 seconds per major iteration. We'd like to
decide on some way to parallelize this across workers. I think the most
fruitful approach will be to do multiple iterations simultaneously. For
example, if you have 10 workers you can do 10 iterations at the same
time, let them synchronize and distribute an updated map, and then go
again. You might want to do >10 iterations before synchronizing to
amortize the cost of doing the synchronization though.

One possible algorithm will be

1. Pick the ~100 cleaning regions with the largest residuals
2. Perform 1 major iteration on each of these regions. These can be
distributed on a first-come first-serve basis to the workers (so that a
slow worker doesn't hold-up the queue). We should not subtract anything
from the maps at this point to prevent race conditions (ie. getting a
different answer based on which worker cleaned which regions).
3. Bring all the clean components back to the master process and do the
clean subtraction.
4. Distribute the updated map to the workers and go again.