fuck, getting two spectral indices from a T-T-T plot is not easy
So let's just explicitly fit for two components from a regular T-T plot.

The problem with fitting a plane to a T-T-T plot is that there is a
non-algebraic relationship between the fitted coefficients and the
spectral indices. Furthermore, I was getting a whole bunch of garbage
when trying to find the zeros numerically.